+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [1.1.0] - 2023-03-13
+
+### Added
+
+- add editor button set viewport resolution
+
+### Changed
+
+- change class_name to add_custom_type
+
 ## [1.0.1] - 2023-03-10
 
 ### Added